==> code/2_analysis/11_random_forest_compare.Rmd <==
---
title: "Random Forest Comparisons"
author: "Leah Lariscy"
output: html_document
---

# Load packages

```{r}
library(tidyverse)
library(here)
library(RColorBrewer)
library(vip)
library(tsibble)
library(ggpubr)
library(yardstick)
```

# Load data

```{r}
#detection frequency metrics summarized
df_metrics <- readRDS(
  here("data/processed_data/rand_forest/cv_metrics_df_sum.rds"))
df_test_metrics <- readRDS(
  here("data/processed_data/final_fit/RF_DF_metrics.rds")) %>% 
  mutate(model = "DF 2")
df_test2_metrics <- readRDS(
  here("data/processed_data/final_fit/bad_fit/RF_DF_metrics.rds")) %>% 
  mutate(model = "DF 2")

#viral load metrics summarized
vl_metrics <- readRDS(
  here("data/processed_data/rand_forest/cv_metrics_vl_sum.rds"))
vl_test_metrics <- readRDS(
  here("data/processed_data/final_fit/RF_VL_metrics.rds")) %>% 
  mutate(model = "VL 2")
vl_test2_metrics <- readRDS(
  here("data/processed_data/final_fit/bad_fit/RF_VL_metrics.rds")) %>% 
  mutate(model = "VL 2")

#all predictors metrics summarized
all_metrics <- readRDS(
  here("data/processed_data/rand_forest/cv_metrics_all_sum.rds"))
all_test_metrics <- readRDS(
  here("data/processed_data/final_fit/RF_ALL_metrics.rds")) %>% 
  mutate(model = "ALL 2")
all_test2_metrics <- readRDS(
  here("data/processed_data/final_fit/bad_fit/RF_ALL_metrics.rds")) %>% 
  mutate(model = "ALL 2")

#bind summarized metrics dataframes
metrics_compare <- rbind(df_metrics,
                         vl_metrics,
                         all_metrics) %>% 
  mutate(config = paste(feature,n,sep = "_"))

test_metrics_compare <- rbind(df_test_metrics,
                              vl_test_metrics,
                              all_test_metrics)

test2_metrics_compare <- rbind(df_test2_metrics,
                              vl_test2_metrics,
                              all_test2_metrics)

saveRDS(metrics_compare,
        here("data/processed_data/rand_forest/metrics_compare.rds"))


#detection frequency n=2 predictions and rsq/rmse
df_n2_train_predictions <- readRDS(here(
  "data/processed_data/rand_forest/df_n2_train_predictions.rds"))

df_n2_test_predictions <- readRDS(here(
  "data/processed_data/rand_forest/df_n2_test_predictions.rds"))
df_n2_test_metrics <- readRDS(here(
  "data/processed_data/rand_forest/df_n2_test_metrics.rds"))


#all predictors n=6 predictions and rsq/rmse
all_n2_train_predictions <- readRDS(here(
  "data/processed_data/rand_forest/all_n2_train_predictions.rds"))

all_n2_test_predictions <- readRDS(here(
  "data/processed_data/rand_forest/all_n2_test_predictions.rds"))
all_n2_test_metrics <- readRDS(here(
  "data/processed_data/rand_forest/all_n2_test_metrics.rds"))

#all predictors n=6 fit (for variable importance viz)
all_n6_train_fit <- readRDS(
  here("data/processed_data/rand_forest/ALL_n6_fit_train.rds"))
df_n6_train_fit <- readRDS(
  here("data/processed_data/rand_forest/DF_n6_fit_train.rds"))
vl_n6_train_fit <- readRDS(
  here("data/processed_data/rand_forest/VL_n6_fit_train.rds"))

#vl predictors n=2 predictions
vl_n2_train_predictions <- readRDS(here(
  "data/processed_data/rand_forest/vl_n2_train_predictions.rds"))

vl_n2_test_predictions <- readRDS(here(
  "data/processed_data/rand_forest/vl_n2_test_predictions.rds"))
vl_n2_test_metrics <- readRDS(here(
  "data/processed_data/rand_forest/vl_n2_test_metrics.rds"))
```

# Visualize

## Make color palette

```{r}
# Extract the last 6 colors from the "Reds", "Greens", and "Blues" palettes
reds <- brewer.pal(9, "Reds")[4:9]
greens <- brewer.pal(9, "Greens")[4:9]
blues <- brewer.pal(9, "Blues")[4:9]

# Combine the colors into a single list
color_palette <- c(reds, greens, blues)
color_palette2 <- c("#238B45", "#08519C", "#CB181D")

# Display the colors to confirm their distinctiveness
print(color_palette)

# Ensure the list has exactly 18 colors
color_palette <- color_palette[1:18]

# Make list for naming models in ggplot
labels <- c("ALL 1","ALL 2","ALL 3","ALL 4","ALL 5","ALL 6",
            "DF 1","DF 2","DF 3","DF 4","DF 5","DF 6",
            "VL 1","VL 2","VL 3","VL 4","VL 5","VL 6")
```

## RMSE

```{r}
rmse_compare <- metrics_compare %>% 
  ggplot(aes(config, avg_rmse, color=config)) +
  geom_point(size = 4) +
    geom_errorbar(aes(ymin = lower_rmse,
                ymax = upper_rmse)) +
  ggthemes::theme_clean() +
  theme(axis.title.x = element_blank(),
        #axis.ticks.x = element_blank(),
        axis.text.x = element_text(angle = 20),
        axis.title.y = element_text(size = 12,face = "bold"),
        legend.text = element_text(size = 10),
        legend.title = element_text(size = 12),
        legend.key.height = unit(0.6, "cm"),
        legend.position = "none") + #no legend after all..
  scale_color_manual(name="Configuration",
                    #labels=labels,
                    values = color_palette) +
  xlab("Feature selection configuration") +
  ylab("RMSE") +
  scale_x_discrete(labels=labels) +
  geom_hline(yintercept = 200, alpha = 0) +
  geom_hline(yintercept = 90, alpha = 0) +
      geom_hline(yintercept = 174,
             linetype = "longdash",
             linewidth = .8,
             color = "red") +
  annotate("text", label="Null model avg", 
           x=16, y=180, size=4, color = "red")

rmse_compare

ggsave(here("figures/random_forest/metrics/rmse_compare.png"), width = 9, height = 5)

saveRDS(rmse_compare, 
        here("figures/random_forest/metrics/rmse_compare.rds"))
```

```{r}
rmse <- metrics_compare %>% 
  select(model,feature,n,avg_rmse,se_rmse) %>% 
  arrange(avg_rmse)

rmse

saveRDS(rmse,
        here("data/processed_data/rand_forest/rmse_compare.rds"))
```

## RSQ

```{r}
rsq_compare <- metrics_compare %>% 
  ggplot(aes(config, avg_rsq, color=config)) +
  geom_point(size = 4) +
    geom_errorbar(aes(ymin = lower_rsq,
                ymax = upper_rsq)) +
  ggthemes::theme_clean() +
  theme(axis.title.x = element_blank(),
        #axis.ticks.x = element_blank(),
        axis.text.x = element_text(angle = 20,
                                   size = 12,
                                   vjust = 0.5),
        axis.text.y = element_text(angle = 15, 
                                   size = 12),
        axis.title = element_text(size = 14,
                                  face = "bold"),
        legend.text = element_text(size = 10),
        legend.title = element_text(size = 12),
        legend.key.height = unit(0.6, "cm"),
        legend.position = "none") + #no legend after all..
  scale_color_manual(name="Configuration",
                    #labels=labels,
                    values = color_palette) +
  #xlab("Feature selection configuration") +
  ylab("R-Squared") +
  scale_x_discrete(labels=labels) +
  geom_hline(yintercept = 0.8, alpha = 0) +
  geom_hline(yintercept = 0.07,
             linetype = "longdash",
             linewidth = .8,
             color = "red") +
  annotate("text", label="Null model avg", 
           x=16, y=.1, size=4, color = "red")

rsq_compare

ggsave(here("figures/random_forest/metrics/rsq_compare.png"), width = 9, height = 5)

saveRDS(rsq_compare,
        here("figures/random_forest/metrics/rsq_compare.rds"))
```

```{r}
rsq <- metrics_compare %>% 
  select(model,feature,n,avg_rsq,se_rsq) %>% 
  arrange(desc(avg_rsq))

rsq

saveRDS(rsq,
        here("data/processed_data/rand_forest/rsq_compare.rds"))
```

### New data

```{r}
#RSQ in early 2022
test_rsq_fig <- test_metrics_compare %>% filter(.metric == "rsq") %>% 
  ggplot(aes(model, .estimate)) +
  geom_point(size = 4, color = color_palette2) +
  ggthemes::theme_clean() +
  theme(axis.title.x = element_blank(),
        #axis.ticks.x = element_blank(),
        #axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 20,
                                   size = 12),
        axis.text.y = element_text(angle = 15,
                                   size = 12),
        axis.title = element_text(size = 14,face = "bold")) +
  ylab("R-Squared") +
  geom_hline(yintercept = .9, alpha = 0) +
  geom_hline(yintercept = .05, alpha = 0)

test_rsq_fig

saveRDS(test_rsq_fig, here("figures/random_forest/metrics/rsq_test_compare.rds"))
```

```{r}
#RSQ in late 2022
test2_rsq_fig <- test2_metrics_compare %>% filter(.metric == "rsq") %>% 
  ggplot(aes(model, .estimate)) +
  geom_point(size = 4, color = color_palette2) +
  ggthemes::theme_clean() +
  theme(axis.title.x = element_blank(),
        #axis.title.y = element_blank(),
        #axis.ticks.x = element_blank(),
        #axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 20,
                                   size = 12),
        axis.text.y = element_text(angle = 15,
                                   size = 12),
        axis.title = element_text(size = 14,face = "bold")) +
  ylab("R-Squared") +
  geom_hline(yintercept = .9, alpha = 0) +
  geom_hline(yintercept = .05, alpha = 0)

test2_rsq_fig

saveRDS(test2_rsq_fig, here("figures/random_forest/metrics/rsq_test2_compare.rds"))
```

## Predictions

### All predictors, n = 2

#### Training data (2020-2021)

```{r}
all_n2_train_fig <- all_n2_train_predictions %>% 
  head(73) %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=cases_predicted), color = "#FB6A4A",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_predicted), color = "#FB6A4A", 
             shape = 17, size = 2) +
  geom_line(aes(y=cases_observed), color = "#003e99",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_observed), color = "#003e99", 
             shape = 16, size = 2) +
      annotate("label", label="RF ALL", 
           x = 18940, y = 1000, size = 5) +
  annotate("label", label="Predicted", x=18650, y=1000, 
           fill="#FB6A4A", size=5, alpha = .7) +
  annotate("label", label="Observed", x=18500, y=1000, 
           fill="#003e99", size=5, alpha = .7) +
  #annotate("text", label=all_n6_train_rsq_round, x=18950, y=3250, size=4) +
  #annotate("text", label="Predicted", x=18550, y=2500, color="#FB6A4A", size=4) +
  #annotate("text", label="Observed", x=18550, y=2650, color="#003e99", size=4) +
  ylab("Cases per week") +
  #xlab("Week") +
  geom_hline(yintercept = 1050, alpha = 0) +
  ggthemes::theme_clean() +
  theme(
        axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 15,
                                   hjust = .9, 
                                   size = 12),
        axis.text.y = element_text(angle = 15,
                                   size = 12),
        axis.title.y = element_text(size = 14)) + 
  scale_x_yearweek(date_breaks = "3 months",
                   date_labels = "%b `%y")

all_n2_train_fig

ggsave(here("figures/random_forest/time_series/all_n2_train.png"))

saveRDS(all_n2_train_fig,
        here("figures/random_forest/time_series/all_n2_train.rds"))
```

#### New data (2022)

```{r}
all_n2_test_fig <- all_n2_test_predictions %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=cases_predicted), color = "#FB6A4A",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_predicted), color = "#FB6A4A", 
             shape = 17, size = 2) +
  geom_line(aes(y=cases_observed), color = "#003e99",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_observed), color = "#003e99",
             size = 2) +
  annotate("label", label="RF ALL", 
           x = 19335, y = 1300, size = 5) +
  annotate("label", label="Predicted", x=19175, y=1300, 
           fill="#FB6A4A", size=5, alpha = .7) +
  annotate("label", label="Observed", x=19075, y=1300, 
           fill="#003e99", size=5, alpha = .7) +
  ylab("Cases per week") +
  #xlab("Week") +
  geom_vline(xintercept = 19113, linetype = 2) +
  geom_hline(yintercept = 1350, alpha = 0) +
  ggthemes::theme_clean() +
  theme(
        axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 15,
                                   hjust = .9,
                                   size = 12),
        axis.text.y = element_text(angle = 15,
                                   size = 12),
        axis.title.y = element_text(size = 14)) +
  scale_x_yearweek(date_breaks = "3 months",
                   date_labels = "%b `%y")

all_n2_test_fig

ggsave(here("figures/random_forest/time_series/all_n2_test.png"))

saveRDS(all_n2_test_fig,
        here("figures/random_forest/time_series/all_n2_test.rds"))
```

### Detection frequency, n = 2

#### Training data (2020-2021)

```{r}
df_n2_train_fig <- df_n2_train_predictions %>% 
  head(73) %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=cases_predicted), color = "#FB6A4A",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_predicted), color = "#FB6A4A", 
             shape = 17, size = 2) +
  geom_line(aes(y=cases_observed), color = "#003e99",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_observed), color = "#003e99", 
             shape = 16, size = 2) +
  annotate("label", label="RF DF", 
           x = 18940, y = 1000, size = 5) +
  #annotate("text", label=all_n6_train_rsq_round, x=18950, y=3250, size=4) +
  #annotate("text", label="Predicted", x=18550, y=2500, color="#FB6A4A", size=4) +
  #annotate("text", label="Observed", x=18550, y=2650, color="#003e99", size=4) +
  ylab("Cases per week") +
  #xlab("Week") +
  geom_hline(yintercept = 1050, alpha = 0) +
  ggthemes::theme_clean() +
  theme(
        axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 15,
                                   hjust = .9,
                                   size = 12),
        axis.text.y = element_text(angle = 15,
                                   size = 12),
        axis.title.y = element_text(size = 14)) + 
  scale_x_yearweek(date_breaks = "3 months",
                   date_labels = "%b `%y")

df_n2_train_fig

ggsave(here("figures/random_forest/time_series/df_n2_train.png"))

saveRDS(df_n2_train_fig,
        here("figures/random_forest/time_series/df_n2_train.rds"))
```

#### New data (2022)

```{r}
df_n2_test_fig <- df_n2_test_predictions %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=cases_predicted), color = "#FB6A4A",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_predicted), color = "#FB6A4A", 
             shape = 17, size = 2) +
  geom_line(aes(y=cases_observed), color = "#003e99",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_observed), color = "#003e99",
             size = 2) +
      annotate("label", label="RF DF", 
           x = 19340, y = 1300, size = 5) +
  #annotate("text", label="Predicted", x=19045, y=1425, color="#FB6A4A", size=4) +
  #annotate("text", label="Observed", x=19045, y=1500, color="#003e99", size=4) +
  ylab("Cases per week") +
  #xlab("Week") +
  geom_vline(xintercept = 19113, linetype = 2) +
  geom_hline(yintercept = 1350, alpha = 0) +
  ggthemes::theme_clean() +
  theme(axis.title.x = element_blank(),
        #axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 15,
                                   hjust = .9,
                                   size = 12),
        axis.text.y = element_text(angle = 15,
                                   size = 12),
        axis.title.y = element_text(size = 14)) +
  scale_x_yearweek(date_breaks = "3 months",
                   date_labels = "%b `%y")

df_n2_test_fig

ggsave(here("figures/random_forest/time_series/df_n2_test.png"))

saveRDS(df_n2_test_fig,
        here("figures/random_forest/time_series/df_n2_test.rds"))
```

### Viral load, n = 2

#### Training data

```{r}
vl_n2_train_fig <- vl_n2_train_predictions %>% 
  head(73) %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=cases_predicted), color = "#FB6A4A",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_predicted), color = "#FB6A4A", 
             shape = 17, size = 2) +
  geom_line(aes(y=cases_observed), color = "#003e99",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_observed), color = "#003e99", 
             shape = 16, size = 2) +
  annotate("label", label="RF VL", 
           x = 18950, y = 1000, size = 5) +
  #annotate("text", label=all_n6_train_rsq_round, x=18950, y=3250, size=4) +
  #annotate("text", label="Predicted", x=18550, y=2500, color="#FB6A4A", size=4) +
  #annotate("text", label="Observed", x=18550, y=2650, color="#003e99", size=4) +
  ylab("Cases per week") +
  #xlab("Week") +
  geom_hline(yintercept = 1050, alpha = 0) +
  ggthemes::theme_clean() +
  theme(
        axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 15,
                                   hjust = .9,
                                   size = 12),
        axis.text.y = element_text(angle = 15,
                                   size = 12),
        axis.title.y = element_text(size = 14)) + 
  scale_x_yearweek(date_breaks = "3 months",
                   date_labels = "%b `%y")

vl_n2_train_fig

ggsave(here("figures/random_forest/time_series/vl_n2_train.png"))

saveRDS(vl_n2_train_fig,
        here("figures/random_forest/time_series/vl_n2_train.rds"))
```

#### New data

```{r}
vl_n2_test_fig <- vl_n2_test_predictions %>% 
  ggplot(aes(x=week)) +
  geom_line(aes(y=cases_predicted), color = "#FB6A4A",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_predicted), color = "#FB6A4A", 
             shape = 17, size = 2) +
  geom_line(aes(y=cases_observed), color = "#003e99",
            linewidth = 2, alpha = .5) +
  geom_point(aes(y=cases_observed), color = "#003e99",
             size = 2) +
  annotate("label", label="RF VL", 
           x = 19340, y = 1300, size = 5) +
  #annotate("text", label="Predicted", x=19045, y=1425, color="#FB6A4A", size=4) +
  #annotate("text", label="Observed", x=19045, y=1500, color="#003e99", size=4) +
  #xlab("Week") +
  geom_vline(xintercept = 19113, linetype = 2) +
  geom_hline(yintercept = 1350, alpha = 0) +
  ggthemes::theme_clean() +
  ylab("Cases per week") +
  theme(axis.title.x = element_blank(),
        #axis.title.y = element_blank(),
        axis.text.x = element_text(angle = 15,
                                   hjust = .9,
                                   size = 12),
        axis.text.y = element_text(angle = 15,
                                   size = 12),
        axis.title.y = element_text(size = 14)) +
  scale_x_yearweek(date_breaks = "3 months",
                   date_labels = "%b `%y")

vl_n2_test_fig 

ggsave(here("figures/random_forest/time_series/vl_n2_test.png"))

saveRDS(vl_n2_test_fig,
        here("figures/random_forest/time_series/vl_n2_test.rds"))
```

## Variable importance

```{r}
all_variables <- c("Plant A N1 VL","Plant A N1 DF","Plant A N2 VL",
                  "Plant A N2 DF","Plant B N1 VL","Plant B N1 DF",
                  "Plant B N2 VL","Plant B N2 DF","Plant C N1 VL",
                  "Plant C N1 DF","Plant C N2 VL","Plant C N2 DF")

redon_pal <- c("#5B859EFF", "#1E395FFF", "#75884BFF", "#1E5A46FF", "#DF8D71FF", "#AF4F2FFF", "#D48F90FF", "#732F30FF", "#AB84A5FF", "#59385CFF", "#D8B847FF", "#B38711FF")

# Extract the last 6 colors from the "Reds", "Greens", and "Blues" palettes
reds <- brewer.pal(9, "Reds")[6:9]
greens <- brewer.pal(9, "Greens")[6:9]
blues <- brewer.pal(9, "Blues")[6:9]

# Combine the colors into a single list
color_palette <- c(reds, greens, blues)

vip::vip(all_n6_train_fit, 
         mapping = aes(fill = .data[["Variable"]]),
         num_features = 12) +
  ggthemes::theme_clean() +
  scale_fill_manual(values = color_palette, labels = all_variables) +
  theme(axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        #legend.position = "none",
        axis.title.x = element_text(size = 12))

ggsave(here("figures/random_forest/variable_importance/vip_all_predictors.png"),
       width = 6, height = 6.5)
```

```{r}
# Combine the colors into a single list
color_palette2 <- c("#EF3B2C", "#A50F15", "#41AB5D", "#006D2C","#4292C6", "#08519C")

vip::vip(df_n6_train_fit, 
         mapping = aes(fill = .data[["Variable"]]),
         num_features = 6) +
  ggthemes::theme_clean() +
  scale_fill_manual(values = color_palette2) +
  theme(axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        #legend.position = "none",
        axis.title.x = element_text(size = 12))

ggsave(here("figures/random_forest/variable_importance/vip_df_predictors.png"),
       width = 6, height = 6.5)
```

```{r}
vip::vip(vl_n6_train_fit, 
         mapping = aes(fill = .data[["Variable"]]),
         num_features = 6) +
  ggthemes::theme_clean() +
  scale_fill_manual(values = color_palette2) +
  theme(axis.text.y = element_blank(),
        axis.ticks.y = element_blank(),
        #legend.position = "none",
        axis.title.x = element_text(size = 12))

ggsave(here("figures/random_forest/variable_importance/vip_vl_predictors.png"),
       width = 6, height = 6.5)
```

```{r}
importance <- vi(all_n6_train_fit) %>% 
  arrange(desc(Importance))

importance
```

```{r}
importance_clean <- importance %>% 
  mutate(plant = str_sub(Variable, 1, 1),
         target = str_sub(Variable, 3, 4),
         data_type = if_else(str_detect(Variable, "POS"), "DF", "VL")) %>% 
  select(!Variable)

importance_clean
```

```{r}
#average importance by plant
plant_vip <- importance_clean %>% 
  group_by(plant) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))

plant_vip

A <- plant_vip$avg[1]
B <- plant_vip$avg[2]
C <- plant_vip$avg[3]

A/B
A/C
B/C
```

```{r}
#average importance by plant and gene
plant_gene_vip <- importance_clean %>% 
  group_by(plant,target) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))

plant_gene_vip
```

```{r}
#average by plant and data type
importance_clean %>% 
  group_by(plant,data_type) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))
```

```{r}
#average by target
gene_vip <- importance_clean %>% 
  group_by(target) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))

gene_vip

N1 <- gene_vip$avg[1]
N2 <- gene_vip$avg[2]

N1/N2
```

```{r}
#average by data type
data_vip <- importance_clean %>% 
  group_by(data_type) %>% 
  summarise(avg = mean(Importance)) %>% 
  arrange(desc(avg))

data_vip

DF <- data_vip$avg[1]
VL <- data_vip$avg[2]

DF/VL
```

```{r}
importance_clean %>% 
  group_by(target,data_type) %>% 
  summarise(sum = sum(Importance)) %>% 
  arrange(desc(sum))
```
